Consolidate API trait conformance tests
